(maint) Update to latest Google Analytics
The Google Analytics property that we currently have is set to be
retired on 1 July 2023.  If you log in, it states:

This property will stop processing data starting 1 July 2023. Use the
GA4 Setup Assistant to complete the next steps.

A new property has been created, and this commit updates to use this
new property.